hotfix(removing-switch-case) refactoring and eslinting
