+++ src/api.js
@@ -1,6 +1,7 @@
 import express from 'express';
 import cors from 'cors';
 import helmet from 'helmet';
+
 import Register from './core/models';
 
 const api = express();
