--- src/core/commandsObject.js
@@ -1,7 +1,0 @@
-import { save, help, remove } from './commands';
-
-export default (params) => ({
-  help: help(params),
-  save: save(params),
-  remove: remove(params),
-});
